Ensemble Techniques Project
Course: Ensemble Techniques
This project is based on the case study of a telecommunication company, which is facing a customer churn issue. The project aims at understanding the pattern of the data and predicting customers who are going to churn based on multiple variables to help the company in retaining their existing customers. The project was accomplished by building a machine learning workflow that will run autonomously with the CSV file and return the best-performing model.